Text_Mining/requirements.txt
streamlit
langchain
PyPDF2
pdfplumber
pdfminer.six
Pillow
pillow_heif
langchain-community
langchain-core
langchain-experimental
langserve[all]
langchain-cli
langsmith
faiss-cpu
unstructured
sentence-transformers
pdf2image
unstructured_inference
pytesseract
PyMuPDF
fitz
frontend
langchain_community
pypdf
langchainhub
langchain-openai
langchain-chroma
streamlit_chat
fastembed
chromadb==0.3.29
python-dotenv
langchain_openai
langchain_groq
llamaapi
langchain_experimental